Refactoring base class
