+++ Lombiq.OrchardCoreApiClient.Tester/Program.cs
@@ -1,3 +1,4 @@
+using Lombiq.OrchardCoreApiClient.Interfaces;
 using Lombiq.OrchardCoreApiClient.Models;
 using System;
 using System.Diagnostics.CodeAnalysis;
@@ -18,7 +19,7 @@ public static class Program
     public static async Task Main(string[] arguments)
     {
         var port = int.TryParse(arguments.FirstOrDefault(), out var customPort) ? customPort : 44335;
-        using var apiClient = new ApiClient(new ApiClientSettings
+        using var apiClient = new ApiClient<IOrchardCoreApi>(new ApiClientSettings
         {
             ClientId = ClientId,
             ClientSecret = ClientSecret,
