+++ Lombiq.OrchardCoreApiClient.Tests.UI/Extensions/TestCaseUITestContextExtensions.cs
@@ -1,4 +1,5 @@
 using Atata;
+using Lombiq.OrchardCoreApiClient.Interfaces;
 using Lombiq.OrchardCoreApiClient.Models;
 using Lombiq.Tests.UI.Extensions;
 using Lombiq.Tests.UI.Services;
@@ -73,7 +74,7 @@ public static class TestCaseUITestContextExtensions
             Category = "UI Test Tenants - Edited",
         };
 
-        using var apiClient = new ApiClient(new ApiClientSettings
+        using var apiClient = new ApiClient<IOrchardCoreApi>(new ApiClientSettings
         {
             ClientId = clientId,
             ClientSecret = clientSecret,
@@ -106,7 +107,7 @@ public static class TestCaseUITestContextExtensions
 
     private static async Task TestTenantCreateAsync(
         UITestContext context,
-        ApiClient apiClient,
+        ApiClient<IOrchardCoreApi> apiClient,
         TenantApiModel createApiModel)
     {
         using (var response = await apiClient.OrchardCoreApi.CreateAsync(createApiModel))
@@ -141,7 +142,7 @@ public static class TestCaseUITestContextExtensions
 
     private static async Task TestTenantSetupAsync(
         UITestContext context,
-        ApiClient apiClient,
+        ApiClient<IOrchardCoreApi> apiClient,
         TenantApiModel createApiModel,
         TenantSetupApiModel setupApiModel)
     {
@@ -159,7 +160,7 @@ public static class TestCaseUITestContextExtensions
 
     private static async Task TestTenantEditAsync(
         UITestContext context,
-        ApiClient apiClient,
+        ApiClient<IOrchardCoreApi> apiClient,
         TenantApiModel editModel,
         TenantSetupApiModel setupApiModel)
     {
@@ -186,7 +187,7 @@ public static class TestCaseUITestContextExtensions
 
     private static async Task TestTenantDisableAsync(
         UITestContext context,
-        ApiClient apiClient,
+        ApiClient<IOrchardCoreApi> apiClient,
         TenantApiModel editModel)
     {
         await apiClient.OrchardCoreApi.DisableAsync(editModel.Name);
@@ -199,7 +200,7 @@ public static class TestCaseUITestContextExtensions
 
     private static async Task TestTenantRemoveAsync(
         UITestContext context,
-        ApiClient apiClient,
+        ApiClient<IOrchardCoreApi> apiClient,
         TenantApiModel editModel)
     {
         await apiClient.OrchardCoreApi.RemoveAsync(editModel.Name);
